Move user logic to UserStore
